changed some doubles to floats

+++ stencil.c
@@ -148,7 +148,7 @@ void stencil(const int nx, const int ny, const int width, const int height,
 void halo_exchange(int rank) {
   for(ii=0; ii < local_nrows; ii++) {
     sendbuf[ii] = subgrid[ii * (local_ncols + 2) + 1];
-  MPI_Sendrecv(sendbuf, local_nrows, MPI_FLOAT, left, tag, recvbuf, local_nrows, MPI_DOUBLE, right, tag, MPI_COMM_WORLD, &status);
+  MPI_Sendrecv(sendbuf, local_nrows, MPI_FLOAT, left, tag, recvbuf, local_nrows, MPI_FLOAT, right, tag, MPI_COMM_WORLD, &status);
   }
   for(ii=0; ii < local_nrows; ii++){
     subgrid[ii * (local_ncols + 2) + local_ncols + 1] = recvbuf[ii];
@@ -156,7 +156,7 @@ void halo_exchange(int rank) {
   /* send to the right, receive from left */
   for(ii=0; ii < local_nrows; ii++){
     sendbuf[ii] = subgrid[ii * (local_ncols + 2) + local_ncols];
-  MPI_Sendrecv(sendbuf, local_nrows, MPI_DOUBLE, right, tag, recvbuf, local_nrows, MPI_DOUBLE, left, tag, MPI_COMM_WORLD, &status);
+  MPI_Sendrecv(sendbuf, local_nrows, MPI_FLOAT, right, tag, recvbuf, local_nrows, MPI_FLOAT, left, tag, MPI_COMM_WORLD, &status);
   }
   for(ii=0; ii < local_nrows; ii++){
     subgrid[ii * (local_ncols + 2)] = recvbuf[ii];
